backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
python-multipart==0.0.9
python-dotenv==1.0.1
langchain==0.2.1
langchain-community==0.2.1
langchain-groq==0.1.3
langchain-huggingface==0.0.1
pypdf==4.2.0
faiss-cpu==1.8.0
pandas>=2.2.2
numpy>=1.26.4
sentence-transformers==3.0.0
docx2txt>=0.8
openpyxl>=3.1.2
unstructured[docx,xlsx]>=0.14.0
groq>=0.9.0